Move tftypes package to root of module.
Rather than making tftypes a namespaced package and having one per
protocol version, we're going to use the same implementation across
protocol versions. To reflect this, move tftypes to the root of the
module, not namespaced by a protocol version, and rewrite the imports
for it.
