+++ tfprotov5/dynamic_value.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"errors"
 
-	"github.com/hashicorp/terraform-plugin-go/tfprotov5/tftypes"
+	"github.com/hashicorp/terraform-plugin-go/tftypes"
 	"github.com/vmihailenco/msgpack"
 )
