+++ tfprotov5/dynamic_value_msgpack.go
@@ -3,7 +3,7 @@ package tfprotov5
 import (
 	"math/big"
 
-	"github.com/hashicorp/terraform-plugin-go/tfprotov5/tftypes"
+	"github.com/hashicorp/terraform-plugin-go/tftypes"
 	"github.com/vmihailenco/msgpack"
 	msgpackCodes "github.com/vmihailenco/msgpack/codes"
 )
